Add SystemContractState
